research_assistant/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-community
langchain-ollama
langchain-huggingface
faiss-cpu
sentence-transformers
cohere
tavily-python
arxiv
pymupdf
python-dotenv
pydantic
langsmith
httpx
ragas
datasets
pandas
